Fixed some 32->64 bit bugs; many more fixes are coming (thanks to B. Watson).
Added commandline options to change almost all properties of a ROM.  Very
useful for developers who constantly change these features but don't want
to re-generate a new stella.pro entry each time.  Thanks again to B. Watson
for original idea and partial implementation.

Fixed problem with events getting 'stuck' when changing states.  For
example, one could be playing a game and, while pressing "Left", exit
the game.  Then, when another game is started, the left direction would
still be selected.  This is a side-effect of Stella initially being a
"one game per launch" application.

For similar reasons as the previous step, fixed problem whereby a
messagebox seen just before ending one game was showing up in the next
game as well.  Still TODO is make sure the sound buffer is cleared
before each new game (sometimes you can hear some sounds from a
previous game).

Further tweaked the '-framerate' commandline argument to override the
defaults when used, but never be saved to the settings file.

Renamed '-altpro' commandline argument back to '-pro', and moved it into
the DEVELOPER_SUPPORT area.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@421 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ stella/src/gui/AudioDialog.cxx
@@ -13,7 +13,7 @@
 // See the file "license" for information on usage and redistribution of
 // this file, and for a DISCLAIMER OF ALL WARRANTIES.
 //
-// $Id: AudioDialog.cxx,v 1.4 2005-05-10 19:20:42 stephena Exp $
+// $Id: AudioDialog.cxx,v 1.5 2005-05-12 18:45:21 stephena Exp $
 //
 //   Based on code from ScummVM - Scumm Interpreter
 //   Copyright (C) 2002-2004 The ScummVM project
@@ -141,7 +141,7 @@ void AudioDialog::saveConfig()
   // Fragsize (requires a restart to take effect)
   i = 1;
   i <<= (myFragsizePopup->getSelectedTag() + 7);
-  if(instance()->settings().getInt("fragsize") != (Int32)i)
+  if(instance()->settings().getInt("fragsize") != i)
   {
     instance()->settings().setInt("fragsize", i);
     restart = true;
